Move generate groups into solver file, consolidate tests
